backend/requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
requests>=2.31.0
python-dotenv>=1.0.0
beautifulsoup4>=4.12.0
uber-rides>=0.6.0
google-auth>=2.23.0
google-auth-oauthlib>=1.1.0
google-auth-httplib2>=0.1.0
google-api-python-client>=2.108.0
pydantic>=2.6.0
